feat🎨: Add registered dataset editor dialog
Add new registered dataset editor dialog component that enables
users to modify dataset metadata, configure time range selection,
and manage SQL queries. The dialog provides comprehensive controls
for dataset name, description, tags, ticker selection, data
intervals, periods, and custom date ranges with SQL query
override capability.
